move request out of entity and extract an interface for usecase
